replace rotate_fail_empty_signers
